Old_version_party-model/party-model/requirementsRTX5060.txt
# ==== requirements.txt ====
torch==2.1.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.1+cu121
numpy>=1.22.0,<2.0.0
opencv-python>=4.6.0,<5.0.0
albumentations>=1.4.18,<2.1.0
tqdm>=4.65.0
scikit-learn>=1.2.0,<1.7.0
transformers>=4.31.0,<4.55.0
PyYAML>=6.0
matplotlib>=3.6.0,<3.10.0
timm>=0.6.12,<1.1.0
#pydensecrf>=1.2.0


